Fix debug format and iface logging macro
Adjusts two debug statements for correctness and consistency: the EMAC timeout log now uses `%u` with explicit unsigned casts to match argument types, and the Ethernet IGMP drop message now uses `NETWORK_IFACE_DEBUG_PUTS` instead of the generic debug macro.

+++ lib-network/src/emac/gd32/emac_eth.cpp
@@ -184,7 +184,7 @@ template <bool T> static void PtpFrameTransmit(uint32_t length) {
             timeout++;
         } while ((0 == tdes0_ttmss_flag) && (timeout < UINT32_MAX));
 
-        EMAC_DEBUG_PRINTF("timeout=%x %d", timeout, (dma_current_txdesc->status & ENET_TDES0_TTMSS));
+        EMAC_DEBUG_PRINTF("timeout=%x %u", static_cast<unsigned>(timeout), static_cast<unsigned>(dma_current_txdesc->status & ENET_TDES0_TTMSS));
 
         dma_current_txdesc->status &= ~ENET_TDES0_TTMSS; ///< Clear timestamp flag
